Legg til støtte for etterfølgende vedtak i behandling og oppdater relaterte DTO-er

+++ src/main/kotlin/no/nav/bidrag/behandling/service/VedtakService.kt
@@ -24,6 +24,7 @@ import no.nav.bidrag.behandling.transformers.vedtak.mapping.fravedtak.tilBeregni
 import no.nav.bidrag.behandling.transformers.vedtak.mapping.tilvedtak.BehandlingTilVedtakMapping
 import no.nav.bidrag.behandling.transformers.vedtak.mapping.tilvedtak.ResultatadBeregningOrkestrering
 import no.nav.bidrag.behandling.transformers.vedtak.validerGrunnlagsreferanser
+import no.nav.bidrag.commons.security.SikkerhetsKontekst
 import no.nav.bidrag.commons.util.secureLogger
 import no.nav.bidrag.domene.enums.behandling.TypeBehandling
 import no.nav.bidrag.domene.enums.vedtak.Beslutningstype
@@ -134,9 +135,9 @@ class VedtakService(
                 .flatMap { vedtaksid ->
                     val opprinneligVedtak = vedtakConsumer.hentVedtak(vedtaksid)!!
                     hentOpprinneligVedtakstidspunkt(opprinneligVedtak)
-                }.toSet() + setOf(PåklagetVedtak(vedtak.vedtaksid.toInt(), vedtak.vedtakstidspunkt!!))
+                }.toSet() + setOf(PåklagetVedtak(vedtak.vedtaksid, vedtak.vedtakstidspunkt!!))
         }
-        return setOf(PåklagetVedtak(vedtak.vedtaksid.toInt(), vedtak.vedtakstidspunkt!!))
+        return setOf(PåklagetVedtak(vedtak.vedtaksid, vedtak.vedtakstidspunkt!!))
     }
 
     @Transactional
